Rename American Express package name
