+++ AmericanExpress/src/main/java/com/braintreepayments/americanexpress/AmericanExpressAnalytics.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.americanexpress
 
 internal object AmericanExpressAnalytics {
