+++ AmericanExpress/src/main/java/com/braintreepayments/americanexpress/AmericanExpressGetRewardsBalanceCallback.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.americanexpress
 
 /**
  * Callback for receiving result of
