+++ AmericanExpress/src/test/java/com/braintreepayments/api/AmericanExpressClientUnitTest.java
@@ -9,6 +9,12 @@ import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
+import com.braintreepayments.americanexpress.AmericanExpressAnalytics;
+import com.braintreepayments.americanexpress.AmericanExpressClient;
+import com.braintreepayments.americanexpress.AmericanExpressGetRewardsBalanceCallback;
+import com.braintreepayments.americanexpress.AmericanExpressResult;
+import com.braintreepayments.americanexpress.AmericanExpressRewardsBalance;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
