+++ AmericanExpress/src/test/java/com/braintreepayments/api/AmericanExpressRewardsBalanceUnitTest.java
@@ -10,6 +10,8 @@ import org.robolectric.RobolectricTestRunner;
 import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertNull;
 
+import com.braintreepayments.americanexpress.AmericanExpressRewardsBalance;
+
 @RunWith(RobolectricTestRunner.class)
 public class AmericanExpressRewardsBalanceUnitTest {
