+++ SharedUtils/src/main/java/com/braintreepayments/api/Json.java
@@ -2,7 +2,7 @@ package com.braintreepayments.api;
 
 import org.json.JSONObject;
 
-class Json {
+public class Json {
 
     /**
      * Returns the value mapped by name if it exists, coercing it if necessary, or fallback if no such mapping exists.
@@ -15,7 +15,7 @@ class Json {
      * @param fallback
      * @return {@link String}
      */
-    static String optString(JSONObject json, String name, String fallback) {
+    public static String optString(JSONObject json, String name, String fallback) {
         if (json == null || json.isNull(name)) {
             return fallback;
         } else {
